+++ R/coca.R
@@ -138,7 +138,7 @@
 #' biplot(bp.sym)                          # produces a Benzecri biplot
 #'
 #' ## extract eigenvalues of the analysis
-#' eigenvals(bp.sym)
+#' vegan::eigenvals(bp.sym)
 #'
 #' ## correlations between beetle and plant score scores on Co-CA axes
 #' corAxis(bp.sym)
@@ -152,7 +152,7 @@
 #' crossval(beetles, plants)
 #' ## so 2 axes are sufficient
 #' ## permutation test to assess significant PLS components - takes a while
-#' bp.perm <- permutest(bp.pred, permutations = 99)
+#' bp.perm <- vegan::permutest(bp.pred, permutations = 99)
 #' bp.perm
 #' }
 #'
@@ -172,7 +172,7 @@
 #' \donttest{
 #' ## determine important PLS components - takes a while
 #' crossval(bryophyte, vascular)
-#' (carp.perm <- permutest(carp.pred, permutations = 99))
+#' (carp.perm <- vegan::permutest(carp.pred, permutations = 99))
 #' }
 #'
 #' ## 2 components again, refit