+++ R/permutest.coca.R
@@ -88,7 +88,7 @@
 #'
 #' ## permutation test
 #' ## (Testing the first 2 axes & only 25 perms for speed.)
-#' bp.perm <- permutest(bp.pred, permutations = 25, n.axes = 2)
+#' bp.perm <- vegan::permutest(bp.pred, permutations = 25, n.axes = 2)
 #' bp.perm
 #'
 #' \dontshow{options(od); do.call(RNGkind, as.list(old_rng))}